Make GlassBlock Colors customizable + Clean up decompiled code
