+++ Ahorn/lang/en_gb.lang
@@ -37,3 +37,7 @@ placements.triggers.SpringCollab2020/RemoveLightSourcesTrigger.tooltips.persiste
 
 # Dash Spring
 placements.entities.SpringCollab2020/dashSpring.tooltips.playerCanUse=Determines whether the player is able to interact with the spring.
+
+# Customizable Glass Block Controller
+placements.entities.SpringCollab2020/CustomizableGlassBlockController.tooltips.starColors=A comma-separated list of all star colors visible in glass blocks in the room.
+placements.entities.SpringCollab2020/CustomizableGlassBlockController.tooltips.bgColor=The background color for glass blocks in the room.
